# be/docs/swagger.json
{
    "schemes": [
        "https",
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "A comprehensive donation and auction management system that transforms donated goods into meaningful impact through transparent auctions and direct donations to institutions in need.",
        "title": "Your Donate Rise API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "yourdonaterise-278016640112.asia-southeast2.run.app",
    "basePath": "/",
    "paths": {
...
        "/donations/final": {
            "get": {
                "description": "Retrieve all items that were directly donated to institutions",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Your Donate Rise API - Final Donations"
                ],
                "summary": "Get all final donations",
                "responses": {
                    "200": {
                        "description": "Final donations fetched successfully",
                        "schema": {
                            "$ref": "#/definitions/utils.SuccessResponseData"
                        }
                    },
                    "400": {
                        "description": "Bad request - Failed to fetch final donations",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/donations/final/user/{user_id}": {
            "get": {
                "description": "Retrieve all final donations made by a specific user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Your Donate Rise API - Final Donations"
                ],
                "summary": "Get final donations by user ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Final donations fetched successfully",
                        "schema": {
                            "$ref": "#/definitions/utils.SuccessResponseData"
                        }
                    },
                    "400": {
                        "description": "Bad request - Invalid user ID or failed to fetch",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "utils.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "example": "error"
                }
            }
        },
        "utils.SuccessResponseData": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}